+++ Assets/Scripts/ObjectDamage.cs
@@ -1,8 +1,6 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class ObjectDamage : MonoBehaviour
 {
-    public float damage;
+    public int damage;
 }
